Moves more code out of cewe2pdf.py into separate files. No logic changes, but renames to get consistent capitalisation of the word keyaccount
